# swagger.yaml
openapi: 3.0.0
info:
  title: ZUCT
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /student/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - STUDENT
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - STUDENT
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/STUDENT"
    post:
      summary: CRUD ACTION update
      tags:
        - STUDENT
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/STUDENT"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/STUDENT"
  /student/findBytutor/{key}:
    get:
      summary: CRUD ACTION findBytutor
      tags:
        - STUDENT
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/STUDENT"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    STUDENT:
      properties:
        _id:
          type: string
        Address:
          type: number
        First_name:
          type: string
        Last_name:
          type: string
        STUDENT_ID:
          type: number
        sex:
          type: string
      required:
        -  _id
        - STUDENT_ID
...